refactor: 重构类型定义，优化导入结构，移除重复定义
